requirements.txt
pyspark==3.3.2
apache-airflow-providers-apache-spark==4.3.0
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-amazon==8.17.0
apache-airflow-providers-apache-kafka==1.4.0
pandas==1.5.3
pyarrow==14.0.1
minio==7.2.7
scikit-learn
psycopg2-binary

